restructuring folder strategy
