+++ velaioProject/src/app/app.component.ts
@@ -1,8 +1,9 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
-import { TaskFormComponent } from './components/task-form/task-form/task-form.component';
-import { TaskListComponent } from './components/task-list/task-list/task-list.component';
+import { TaskFormComponent } from './components/task-form/task-form.component';
+import { TaskListComponent } from './components/task-list/task-list.component';
+
 
 
 @Component({
