+++ velaioProject/src/app/components/task-list/task-list.component.ts
@@ -1,7 +1,8 @@
 import { Component, inject } from '@angular/core';
 
 import { CommonModule } from '@angular/common';
-import { TaskService } from '../../../service/task.service';
+import { TaskService } from '../../service/task.service';
+
 
 @Component({
   selector: 'app-task-list',
